#  optimize Varint paring by avx instruction. In our funciton level ben…