…chmarking, we get about 2.7x  maximum performance

 boost.
 un on (112 X 3100 MHz CPU s)
 CPU Caches:
 L1 Data 48K (x56)
 L1 Instruction 32K (x56)
 L2 Unified 1280K (x56)
 L3 Unified 43008K (x2)
 Load Average: 5.27, 2.01, 1.00
 ***WARNING*** CPU scaling is enabled, the benchmark real time measurements may be noisy and will incur extra overhead.
 ---------------------------------------------------------------------------
 Benchmark                 Time             CPU   Iterations UserCounters...
 ---------------------------------------------------------------------------
 BM_Parse32_1           4.89 ns         4.88 ns    125973673 bytes_per_second=195.237M/s
 BM_ParseAvx32_1        4.96 ns         4.95 ns    143077275 bytes_per_second=192.53M/s
 BM_Parse32_2           6.38 ns         6.38 ns    110719486 bytes_per_second=298.887M/s
 BM_ParseAvx32_2        5.89 ns         5.89 ns    109395587 bytes_per_second=323.852M/s
 BM_Parse32_3           15.4 ns         15.4 ns     45830635 bytes_per_second=185.982M/s
 BM_ParseAvx32_3        13.1 ns         13.1 ns     53303984 bytes_per_second=218.311M/s
 BM_Parse32_4           17.7 ns         17.7 ns     39587623 bytes_per_second=215.326M/s
 BM_ParseAvx32_4        13.0 ns         13.0 ns     55327722 bytes_per_second=294.205M/s
 BM_Parse32_5           20.7 ns         20.7 ns     33756288 bytes_per_second=230.342M/s
 BM_ParseAvx32_5        13.5 ns         13.5 ns     55914882 bytes_per_second=353.85M/s
 BM_Parse32_6           23.6 ns         23.6 ns     31853344 bytes_per_second=242.756M/s
 BM_ParseAvx32_6        12.8 ns         12.8 ns     58112963 bytes_per_second=446.48M/s
 BM_Parse32_7           25.9 ns         25.9 ns     27330376 bytes_per_second=257.454M/s
 BM_ParseAvx32_7        11.7 ns         11.7 ns     61916955 bytes_per_second=571.664M/s
 BM_Parse32_8           25.5 ns         25.5 ns     27485758 bytes_per_second=298.895M/s
 BM_ParseAvx32_8        10.9 ns         10.9 ns     64392599 bytes_per_second=699.49M/s
 BM_Parse32_9           27.3 ns         27.3 ns     25693587 bytes_per_second=314.493M/s
 BM_ParseAvx32_9        10.6 ns         10.6 ns     64708421 bytes_per_second=806.913M/s
 BM_Parse32_10          32.8 ns         32.8 ns     23923442 bytes_per_second=290.943M/s
 BM_ParseAvx32_10       12.4 ns         12.4 ns     63656359 bytes_per_second=768.539M/s
 BM_Parse64_1           5.31 ns         5.31 ns    100000000 bytes_per_second=179.756M/s
 BM_ParseAvx64_1        5.57 ns         5.57 ns    133550234 bytes_per_second=171.369M/s
 BM_Parse64_2           6.36 ns         6.35 ns    111933946 bytes_per_second=300.137M/s
 BM_ParseAvx64_2        6.32 ns         6.32 ns    114442784 bytes_per_second=302.015M/s
 BM_Parse64_3           15.4 ns         15.4 ns     45701054 bytes_per_second=185.904M/s
 BM_ParseAvx64_3        11.5 ns         11.5 ns     60860685 bytes_per_second=249.099M/s
 BM_Parse64_4           18.7 ns         18.7 ns     37447788 bytes_per_second=204.247M/s
 BM_ParseAvx64_4        11.6 ns         11.6 ns     62635534 bytes_per_second=329.815M/s
 BM_Parse64_5           21.1 ns         21.1 ns     33306121 bytes_per_second=226.324M/s
 BM_ParseAvx64_5        11.5 ns         11.5 ns     59381621 bytes_per_second=413.83M/s
 BM_Parse64_6           23.7 ns         23.7 ns     29585924 bytes_per_second=241.676M/s
 BM_ParseAvx64_6        11.5 ns         11.5 ns     59528160 bytes_per_second=497.383M/s
 BM_Parse63_7           28.4 ns         28.4 ns     26271316 bytes_per_second=235.006M/s
 BM_ParseAvx64_7        13.4 ns         13.4 ns     48191809 bytes_per_second=497.948M/s
 BM_Parse64_8           33.5 ns         33.5 ns     23452755 bytes_per_second=228.007M/s
 BM_ParseAvx64_8        16.4 ns         16.4 ns     44373249 bytes_per_second=464.317M/s
 BM_Parse64_9           32.1 ns         32.1 ns     20919866 bytes_per_second=267.413M/s
 BM_ParseAvx64_9        15.6 ns         15.6 ns     43917623 bytes_per_second=551.885M/s
 BM_Parse64_10          34.8 ns         34.8 ns     20118905 bytes_per_second=274.119M/s
 BM_ParseAvx64_10       15.5 ns         15.5 ns     44864739 bytes_per_second=613.729M/s

in our object level test. End to End benchmarking for object that all are varint fields ,we can get about 20% performance boost